requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
humanize==4.9.0
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.1
urllib3==2.2.3
Werkzeug==3.1.2
WTForms==3.2.1
